+++ cli/CHANGELOG.md
@@ -9,6 +9,7 @@
 - CLI performs git pull on each subdirectory repository from a workspace directory
 - CLI provides interactive TUI for git pulls
 - CLI git interactive mode launches browser with GitHub compare page
+- CLI git prints repo state
 
 [Unreleased]: https://github.com/eighty4/maestro/compare/maestro-v0.2.0...HEAD
 [0.2.0]: https://github.com/eighty4/maestro/releases/tag/maestro-v0.2.0