+++ git/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Sync API parallelizes git pulls across multiple git repositories
 - Sync API provides RemoteHost for accessing featurs of the remote's Git services such as GitHub's ref compare page
+- Sync API collects local change and stash counts in repository
 
 [Unreleased]: https://github.com/eighty4/maestro/compare/maestro_git-v0.2.0...HEAD
 [0.2.0]: https://github.com/eighty4/maestro/releases/tag/maestro_git-v0.2.0